remove kaliedo common structure file
